rename news to standard
