+++ src/components/standard/articleHeadline.tsx
